Fix default cleanup state machine
