--- content/automate/ManageIQ/Transformation/Common.class/__methods__/assesstransformationcleanup.yaml
@@ -1,13 +1,0 @@
----
-object_type: method
-version: 1.0
-object:
-  attributes:
-    name: AssessTransformationCleanup
-    display_name: 
-    description: 
-    scope: instance
-    language: ruby
-    location: inline
-    options: {}
-  inputs: []
