+++ content/automate/ManageIQ/Transformation/StateMachines/VMTransformation.class/posttransform.yaml
@@ -25,7 +25,7 @@ object:
       on_error: /System/CommonMethods/MiqAe.WeightedUpdateStatus(weight => 5, description
         => "Power-on VM", task_message => "Migrating")
   - State8:
-      value: "/Transformation/Infrastructure/VM/${state_var#destination_ems_type}/CheckPoweredOn"
+      value: "/Transformation/Infrastructure/VM/Common/CheckPoweredOn"
       on_entry: /System/CommonMethods/MiqAe.WeightedUpdateStatus(weight => 40, description
         => "Power-on VM", task_message => "Migrating")
       on_exit: /System/CommonMethods/MiqAe.WeightedUpdateStatus(weight => 40, description
